requirements.txt
# Essential dependencies only
gspread>=6.0.0
google-auth>=2.23.0
requests>=2.31.0
beautifulsoup4>=4.12.0
